+++ app/lib/payment/gateways/stripe/billing/unique_payment.rb
@@ -13,7 +13,7 @@ module Payment
                                            })
           rescue ::Stripe::CardError => e
             charge = ::Stripe::Charge.retrieve(e.error.payment_intent.latest_charge)
-            raise CardErrors.new(
+            raise Stripe::CardErrors.new(
               external_id: charge.payment_intent,
               code: e.code,
               message: e.message,
